refactored code in expenses and cos results registration
