+++ client/src/pages/CostOfSalesResults/CostOfSalesResultsRegistration.tsx
@@ -18,7 +18,7 @@ import {
 import { filterCostOfSaleResults } from '../../api/CostOfSalesResultsEndpoint/FilterCostOfSalesResults'
 import { createCostOfSaleResults } from '../../api/CostOfSalesResultsEndpoint/CreateCostOfSalesResults'
 import { overwriteCostOfSaleResults } from '../../api/CostOfSalesResultsEndpoint/OverwriteCostOfSalesResults'
-import {handleDisableKeysOnNumberInputs, formatNumberWithCommas, removeCommas, sortByFinancialYear} from '../../utils/helperFunctionsUtil' 
+import {handleDisableKeysOnNumberInputs, formatNumberWithCommas, removeCommas, sortByFinancialYear, handleGeneralResultsInputChange} from '../../utils/helperFunctionsUtil' 
 import { getCostOfSale } from '../../api/CostOfSalesEndpoint/GetCostOfSale'
 import { MAX_NUMBER_LENGTH } from '../../constants'
 
@@ -165,52 +165,8 @@ const CostOfSalesResultsRegistration = () => {
   }
 
   const handleChange = (index, event) => {
-   
-    const { name, value } = event.target
-    // Remove commas to get the raw number
-    // EG. 999,999 → 999999 in the DB
-    const rawValue = removeCommas(value)
-
     const nonFinancialValuesArray = ['year', 'month'] 
-
-    if (!nonFinancialValuesArray.includes(name)) {
-      if (rawValue.length > MAX_NUMBER_LENGTH) {
-        return
-      }
-    }
-
-    setFormData((prevFormData) => {
-      return prevFormData.map((form, i) => {
-        if (i === index) {
-          const resetFields = {
-            params: ['months'],
-          }
-          let month = form.month
-          if (name == 'year' && value == '') {
-            form.month = ''
-            setFilteredMonth((prev) => {
-              return prev.map((eachMonth, monthIndex) => {
-                if (index == monthIndex) {
-                  return [{}]
-                }
-                return eachMonth
-              })
-            })
-          }
-          const fieldsToReset = resetFields[name] || []
-          const resetValues = fieldsToReset.reduce((acc, field) => {
-            acc[field] = ''
-            return acc
-          }, {})
-          return {
-            ...form,
-            [name]: rawValue,
-            ...resetValues,
-          }
-        }
-        return form
-      })
-    })
+    handleGeneralResultsInputChange(index, event, setFormData, nonFinancialValuesArray, setFilteredMonth)
   }
